Instead of using a separate component for each category, a generic category component was created
