+++ src/Root.tsx
@@ -1,9 +1,8 @@
 import { HashRouter as Router, Routes, Route } from 'react-router-dom';
 import App from './App';
-import { HomePage, NotFoundPage, PhonePage } from './pages';
-import TabletsPage from './components/TabletsPage/TabletsPage';
-import AccessoriesPage from './components/AccessoriesPage/AccessoriesPage';
+import { HomePage, NotFoundPage } from './pages';
 import CartPage from './pages/Cart/CartPage';
+import CategoryPage from './components/CategoryPage/CategoryPage';
 
 export const Root = () => {
   return (
@@ -11,9 +10,9 @@ export const Root = () => {
       <Routes>
         <Route path="/" element={<App />}>
           <Route index element={<HomePage />} />
-          <Route path="/phones" element={<PhonePage />} />
-          <Route path="/tablets" element={<TabletsPage />} />
-          <Route path="/accessories" element={<AccessoriesPage />} />
+          <Route path="/phones" element={<CategoryPage />} />
+          <Route path="/tablets" element={<CategoryPage />} />
+          <Route path="/accessories" element={<CategoryPage />} />
           <Route path="/cart" element={<CartPage />} />
           <Route path="*" element={<NotFoundPage />} />
         </Route>
